pyproject.toml
[tool.poetry]
name = "airbus_ship_detection"
version = "0.1.0"
description = ""
authors = ["Nazar Polishchuk <nazarr.polishchuk@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.23.1"
pandas = "^1.4.3"
python-dotenv = "^0.20.0"
dvc = "^2.17.0"

[tool.poetry.dev-dependencies]
notebook = "^6.4.12"
matplotlib = "^3.5.3"
scikit-learn = "^1.1.2"
seaborn = "^0.11.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
